app/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
python-multipart
scikit-learn
matplotlib
seaborn
numpy
pre-commit

